\name{catenary-class}
\alias{catenary-class}
\title{Create a class for catenary}
\value{
  an object of class \code{catenary}
}
\description{
  Creates a basic catenary with endpoints
}
\section{Slots}{
  \itemize{ \item c1: shape parameter \item c2: x-location
  parameter \item lambda: y-location parameter \item
  endpoints: left and right endpoint in data frame \item L:
  length of caternary }
}
\examples{
getSlots("catenary")
}

