% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctesiphon.R
\docType{data}
\name{ctesiphon}
\alias{ctesiphon}
\title{Coordinates of external and iternal edge of the Ctesiphon}
\format{A list with two data frames}
\description{
Using an image of the Ctesiphon, the x and y coordinates of
the internal and external edge were recorded. The original image 
and the image with the points overlaid can be found in the inst folder
}
\details{
\itemize{
  \item internal dataframe with x and y point for internal part of arch
  \item external dataframe with x and y point for external part of arch
}
}
\keyword{datasets}

