% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\docType{methods}
\name{getPoints}
\alias{getPoints}
\alias{getPoints}
\alias{getPoints,catenary-method}
\title{Method to get points for catenary}
\value{
data frame of points
}
\description{
Gives points for plotting
}
\keyword{internal}
