% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedCatenary.R
\docType{methods}
\name{gof}
\alias{gof}
\alias{gof}
\alias{gof,catenary-method}
\title{Get goodness of fit statistics for fitted catenary}
\arguments{
\item{x}{A \code{fittedCatenary} object}
}
\value{
data frame of summary statistics
}
\description{
Returns data frame of summary stats for models
}
