% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\docType{methods}
\name{L}
\alias{L}
\alias{L}
\alias{L,catenary-method}
\title{Get length for catenary}
\arguments{
\item{x}{A \code{catenary} object}
}
\value{
length
}
\description{
Returns the length of catenary
}
\examples{
tmp <- catenary(c1=1,c2=2,lambda=3,x0=0,x1=3)
L(tmp)
}
