% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\docType{methods}
\name{minmax}
\alias{minmax}
\alias{minmax,catenary-method}
\title{Method to get min and max of catenary}
\arguments{
\item{object}{a catenary object}
}
\value{
gives min or max values of catenary
}
\description{
Gives min or max
}
\examples{
cat <- catenary(c1=1,c2=1,lambda=1)
minmax(cat)
}

