% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmaxLength.R
\name{minmaxLength}
\alias{minmaxLength}
\title{Given two endpoints gives max and min length of catenary joining them}
\usage{
minmaxLength(endpoints)
}
\arguments{
\item{endpoints}{dataframe with x and y column}
}
\value{
vector of min and max length
}
\description{
Takes data frame giving endpoints and return min length and max
length before hits the ground
}
\note{
February 11 2013
}
\examples{
x <- c(-1,1)
y <- c(2,2)
endpoints <- data.frame(x=x,y=y)
minmaxLength(endpoints)
}
\author{
Jono Tuke <simon.tuke@adelaide.edu.au>
}
\keyword{internal}
