% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R, R/fittedCatenary.R
\docType{methods}
\name{show}
\alias{show}
\alias{show}
\alias{show,}
\alias{catenary-method}
\alias{show}
\alias{show,fittedCatenary-method}
\title{summary function}
\arguments{
\item{x}{catenary object}

\item{...}{extra}

\item{na.rm}{boolean to remove NAs}

\item{object}{fittedCatenary object}
}
\description{
gives pretty summary

show method for fitted Catenary
}
\author{
Jono Tuke, Matthew Roughan
}
