% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\docType{methods}
\name{vertex}
\alias{vertex}
\alias{vertex,catenary-method}
\title{Method to get vertex for catenary}
\arguments{
\item{object}{a catenary object}
}
\value{
coordinates of vertex
}
\description{
Gives vertex point
}
\examples{
cat <- catenary(c1=1,c2=1,lambda=1)
vertex(cat)
}

