% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R
\name{catenary-class}
\alias{catenary-class}
\title{Create a class for catenary}
\value{
an object of class \code{catenary}
}
\description{
Creates a basic catenary with endpoints
}
\section{Slots}{

\describe{
\item{\code{c1}}{shape parameter}

\item{\code{c2}}{x-location parameter}

\item{\code{lambda}}{y-location parameter}

\item{\code{endpoints}}{left and right endpoint in data frame}

\item{\code{L}}{length of catenary}
}}
\examples{
getSlots("catenary")
}

