% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R, R/fittedCatenary.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,catenary-method}
\alias{plot,fittedCatenary-method}
\title{Set generic plot}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{...}{extra}

\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{fit}{type of fit to show at present two choices "cat" and "para"}

\item{envelope}{type of envelope to show at present two choices "cat" and "para"}
}
\description{
Overload plot

Generic plot

Method that can plot fits and function envelopes
}
\examples{
tmp <- catenary(c1=1,c2=3,lambda=1,x0=0,x1=4)
plot(tmp)
}
\author{
Jono Tuke, Matthew Roughan
}

