% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catenary.R, R/fittedCatenary.R
\name{Summary}
\alias{Summary}
\alias{Summary,catenary-method}
\alias{Summary,fittedCatenary-method}
\title{summary function}
\arguments{
\item{x}{catenary object}

\item{...}{extra}

\item{na.rm}{boolean to remove NAs}
}
\description{
gives pretty summary

Nicely presented summary
}
\author{
Jono Tuke, Matthew Roughan
}

